LabelsDetection/requirements.txt
Flask
Werkzeug
python-dotenv
boto3
python-magic
libmagic
requests
flask-cors

